Move action context into XrUtility
