+++ samples/BasicXrApp/BasicXrApp_uwp.vcxproj.filters
@@ -48,5 +48,6 @@
       <Filter>Assets</Filter>
     </None>
     <None Include="packages.config" />
+    <None Include="$(OpenXRLoaderBinaryRoot)\bin\openxr_loader.dll" />
   </ItemGroup>
 </Project>
